pyproject.toml
[tool.poetry]
name = "llama-index-sandbox"
version = "0.1.0"
description = "Sandbox project for testing out various llama-index functionalities."
authors = ["Victor I. Afolabi <javafolabi@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "sandbox", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.10.30"
llama-index-core = "^0.10.30"
llama-index-readers-file = "^0.1.19"
llama-index-llms-ollama = "^0.1.2"
llama-index-embeddings-huggingface = "^0.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pydocstyle = { version = "^6.3.0", extras = ["tomli"] }
docformatter = { version = "^1.7.5", extras = ["tomli"] }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Docformatter.
[tool.docformatter]
in-place = true
recursive = true
blank = true
close-quotes-on-newline = true


# Style hinting.
[tool.mypy]
files = ["src/", "examples/"]
exclude = ["tests/", "venv/", ".venv/"]
namespace_packages = true
explicit_package_bases = true
strict = true
check_untyped_defs = true
ignore_missing_imports = true
disallow_any_generics = true
disallow_incomplete_defs = true
implicit_reexport = false
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
# show_error_content = true
warn_unused_configs = true
warn_unused_ignores = true
# warn_redundant_cast = true
warn_return_any = true
warn_unreachable = true
disallow_subclassing_any = false
disallow_untyped_decorators = false


[tool.pydocstyle]
convention = "google"
add_ignore = [
  # Missing docstring in public module.
  "D100",
  # Missing docstring in public class.
  "D101",
  # Missing docstring in public package.
  "D104",
  # No blank line allowed after function docstring.
  "D202",
  # 1 blank line required between summary line and description.
  "D205",
  # First line should end with a period, question mark, or exclamation point.
  "D415",
]


[tool.ruff.format]
quote-style = "single"
